chore: rename class to `ElasticsearchBackend`
