+++ tests/server/datasets/test_dao.py
@@ -20,7 +20,7 @@ from rubrix.server.apis.v0.models.commons.model import TaskType
 from rubrix.server.apis.v0.models.datasets import DatasetDB
 from rubrix.server.daos.datasets import DatasetsDAO
 from rubrix.server.daos.records import dataset_records_dao
-from rubrix.server.elasticseach.client_wrapper import create_es_wrapper
+from rubrix.server.elasticseach.backend import create_es_wrapper
 from rubrix.server.elasticseach.mappings.text_classification import (
     text_classification_mappings,
 )
